requirements.txt
flask
langchain
langchain_community
langchain_openai
python-dotenv
openai
pydantic
gunicorn